refactor(telegram): move TelegramSender to its own module
Breaks the static import cycle sentrux flagged between
``dev_admin.py`` and ``handlers.py``. Both modules now share the
same ``TelegramSender`` dataclass via a new ``sender.py`` module
in the package, so no parallel ``Protocol`` is needed and no
``TYPE_CHECKING`` dance is required.

``handlers.py`` re-exports ``TelegramSender`` so every existing
caller (``bot.py``, ``model_picker_runtime``, the test suite)
keeps the historical ``from app.integrations.telegram.handlers
import TelegramSender`` import path working unchanged.

Follow-up opportunity (out of scope): ``model_picker.py``,
``status.py``, and ``lcm_status.py`` each declare their own
``(_)TelegramSenderLike`` ``Protocol`` for the same reason. With
the shared module in place those can fold back to the real type.
